Replace old form with new one
