+++ src/core/dashboard/calendar/EventList.tsx
@@ -8,8 +8,8 @@ import { useToast } from '~/components/ui/useToast';
 import { type RefObject, useEffect } from 'react';
 import Event from './Event';
 import { Button } from '~/components/ui/Button';
-import { useNewEvent } from './new-event/newEventStore';
 import { useDragScroll } from './useDragScroll';
+import { useNewEvent } from './new-event/NewEvent';
 
 export default function EventList() {
   const currentDay = useCalendar(s => s.currentDay);
