+++ src/core/dashboard/calendar/new-event/Briefing.tsx
