+++ src/core/dashboard/calendar/new-event/NewEvent.tsx
@@ -1,16 +1,24 @@
 import { Button } from '~/components/ui/Button';
 import { Sheet, SheetContent, SheetTrigger } from '~/components/ui/Sheet';
-import { useState } from 'react';
 import MultiStepForm from '~/components/MultiStepForm';
-import StepOne, { stepOneSchema } from '../new-event-2/Step1';
-import StepTwo, { stepTwoSchema } from '../new-event-2/Step2';
-import StepThree, { stepThreeSchema } from '../new-event-2/Step3';
+import StepOne, { stepOneSchema } from './Step1';
+import StepTwo, { stepTwoSchema } from './Step2';
+import StepThree, { stepThreeSchema } from './Step3';
 import { CalendarPlusIcon } from 'lucide-react';
+import { create } from 'zustand';
+
+export const useNewEvent = create<{
+  sheetOpened: boolean;
+  setSheetOpened: (open: boolean) => void;
+}>()(set => ({
+  sheetOpened: false,
+  setSheetOpened: open => set(() => ({ sheetOpened: open })),
+}));
 
 export const formSchema = stepOneSchema.and(stepTwoSchema).and(stepThreeSchema);
 
 export default function NewEvent() {
-  const [sheetOpened, setSheetOpened] = useState(false);
+  const { sheetOpened, setSheetOpened } = useNewEvent();
 
   return (
     <Sheet open={sheetOpened} onOpenChange={setSheetOpened}>
@@ -18,7 +26,7 @@ export default function NewEvent() {
         <Button
           variant='fab'
           size='fab'
-          className='fixed bottom-24 right-4'
+          className='fixed bottom-24 right-4 lg:hidden'
           aria-label='Create event'
         >
           <CalendarPlusIcon />
